section7_kserve/4_transformer/requirements.txt
kserve==0.9
pillow==9.0.0
argparse>=1.4.0